# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/appointment/create_appointment": {
      "post": {
        "operationId": "AppointementController_create_appointment",
        "summary": "Doctor creates appointment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Created Appointment"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "Invalid_id": {
                    "value": {
                      "statusCode": 400,
                      "message": "Id is not in valid format",
                      "error": "Bad Request"
                    }
                  },
                  "Date_format": {
                    "value": {
                      "statusCode": 400,
                      "message": [
                        "start_date and end_date must match /^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}$/ regular expression"
                      ],
                      "error": "Bad Request"
                    }
                  },
                  "Unable_to_cast_date_type": {
                    "value": {
                      "statusCode": 400,
                      "message": "start_date and/or end_date are not a correct time",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Appointment"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointment/{appointment_id}/confirm_appointment": {
      "patch": {
        "operationId": "AppointementController_confirm_appointment",
        "summary": "Patient confirm appoinment",
        "parameters": [
          {
            "name": "appointment_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Appointment confirmed"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "Invalid_id": {
                    "value": {
                      "statusCode": 400,
                      "message": "Id is not in valid format",
                      "error": "Bad Request"
                    }
                  },
                  "No_appointment_for_this_user": {
                    "value": {
                      "statusCode": 400,
                      "message": "No appointment for this patient",
                      "error": "Bad Request"
                    }
                  },
                  "Wrong_status": {
                    "value": {
                      "statusCode": 400,
                      "message": "Appointment is not in the CREATED status",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Appointment"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointment/{appointment_id}/cancel_appointment": {
      "patch": {
        "operationId": "AppointementController_cancel_appointment",
        "summary": "Patient/Doctor cancel the appoinment",
        "parameters": [
          {
            "name": "appointment_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Appointment canceled"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "Invalid_id": {
                    "value": {
                      "statusCode": 400,
                      "message": "Id is not in valid format",
                      "error": "Bad Request"
                    }
                  },
                  "No_appointment_for_this_user": {
                    "value": {
                      "statusCode": 400,
                      "message": "No appointment for this patient",
                      "error": "Bad Request"
                    }
                  },
                  "Wrong_status": {
                    "value": {
                      "statusCode": 400,
                      "message": "Appointment is not in the CONFIRMED or CREATED status",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Appointment"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointment/{appointment_id}/complete_appointment": {
      "patch": {
        "operationId": "AppointementController_complete_appointment",
        "summary": "Marks the appointment completed and adds Payment Info",
        "parameters": [
          {
            "name": "appointment_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentInfoDto"
              },
              "examples": {
                "example1": {
                  "value": {
                    "amount": 0,
                    "status": "string",
                    "method": "string",
                    "date": "2023-11-07T12:30:00"
                  },
                  "summary": "An example appointment request "
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Appointment completed with payment info"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "Invalid_id": {
                    "value": {
                      "statusCode": 400,
                      "message": "Id is not in valid format",
                      "error": "Bad Request"
                    }
                  },
                  "No_appointment_for_this_user": {
                    "value": {
                      "statusCode": 400,
                      "message": "No appointment for this doctor",
                      "error": "Bad Request"
                    }
                  },
                  "Must_be_CONFIRMED": {
                    "value": {
                      "statusCode": 400,
                      "message": "Appointment is not in the CONFIRMED status",
                      "error": "Bad Request"
                    }
                  },
                  "Date_format": {
                    "value": {
                      "statusCode": 400,
                      "message": [
                        "date must match /^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}$/ regular expression"
                      ],
                      "error": "Bad Request"
                    }
                  },
                  "Unable_to_cast_date_type": {
                    "value": {
                      "statusCode": 400,
                      "message": "date are not a correct time",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Appointment"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointment/doctor/appointment": {
      "get": {
        "operationId": "AppointementController_get_doctor_appointment",
        "summary": "Gets all the appointment for a doctor",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "docs": [
                      {
                        "_id": "string",
                        "patient_id": "string",
                        "doctor_id": "string",
                        "start_date": "2023-12-07T12:30:00.000Z",
                        "end_date": "2023-12-07T12:50:00.000Z",
                        "status": [
                          "CREATED"
                        ],
                        "__v": 0
                      },
                      {
                        "_id": "string",
                        "patient_id": "string",
                        "title": "Cool appointment",
                        "doctor_id": "string",
                        "start_date": "2023-11-07T10:30:00.000Z",
                        "end_date": "2023-11-07T10:40:00.000Z",
                        "status": [
                          "CREATED"
                        ],
                        "__v": 0
                      },
                      {
                        "_id": "string",
                        "patient_id": "string",
                        "title": "title",
                        "doctor_id": "string",
                        "start_date": "2023-11-07T10:30:00.000Z",
                        "end_date": "2023-11-07T11:30:00.000Z",
                        "status": [
                          "CONFIRMED"
                        ],
                        "__v": 0
                      }
                    ],
                    "totalDocs": 3,
                    "limit": 25,
                    "totalPages": 1,
                    "page": 1,
                    "pagingCounter": 1,
                    "hasPrevPage": false,
                    "hasNextPage": false,
                    "prevPage": null,
                    "nextPage": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Appointment"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/appointment/patient/appointment": {
      "get": {
        "operationId": "AppointementController_get_patient_appointment",
        "summary": "Gets all the appointment for a patient",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "docs": [
                      {
                        "_id": "string",
                        "patient_id": "string",
                        "doctor_id": "string",
                        "start_date": "2023-12-07T12:30:00.000Z",
                        "end_date": "2023-12-07T12:50:00.000Z",
                        "status": [
                          "CREATED"
                        ],
                        "__v": 0
                      },
                      {
                        "_id": "string",
                        "patient_id": "string",
                        "title": "Cool appointment",
                        "doctor_id": "string",
                        "start_date": "2023-11-07T10:30:00.000Z",
                        "end_date": "2023-11-07T10:40:00.000Z",
                        "status": [
                          "CREATED"
                        ],
                        "__v": 0
                      },
                      {
                        "_id": "string",
                        "patient_id": "string",
                        "title": "title",
                        "doctor_id": "string",
                        "start_date": "2023-11-07T10:30:00.000Z",
                        "end_date": "2023-11-07T11:30:00.000Z",
                        "status": [
                          "CONFIRMED"
                        ],
                        "__v": 0
                      }
                    ],
                    "totalDocs": 3,
                    "limit": 25,
                    "totalPages": 1,
                    "page": 1,
                    "pagingCounter": 1,
                    "hasPrevPage": false,
                    "hasNextPage": false,
                    "prevPage": null,
                    "nextPage": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Appointment"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/appointment/{appointment_id}/edit_payment_info": {
      "patch": {
        "operationId": "AppointementController_edit_payment_info",
        "summary": "Edit Payment Info",
        "parameters": [
          {
            "name": "appointment_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditAmountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Payment info for appointment has been  edited"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Appointment Not Found",
                    "error": "Bad Request"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Appointment"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Theraline endpoints",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateAppointmentDto": {
        "type": "object",
        "properties": {
          "patient_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "start_date": {
            "format": "date-time",
            "type": "string"
          },
          "end_date": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "patient_id",
          "title",
          "start_date",
          "end_date"
        ]
      },
      "PaymentInfoDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "status",
          "method",
          "date"
        ]
      },
      "EditAmountDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "status"
        ]
      },
...
    }
  }
}